Does it matter? HA and DR in Serverless solutions.
When we talk to customers about Serverless architectures the topic of High Availability and Disaster Recovery comes up, the perception usually is that in a Serverless world we don’t care about these because, well; we are Serverless! 
The flip side of the coin is different. In this talk, I will discuss the design and planning considerations for building robust serverless solutions.